chore: more split
