+++ app/routes/caption-editor/watch.tsx
@@ -1,4 +1,4 @@
-import { tinyassert, uniqBy } from "@hiogawa/utils";
+import { uniqBy } from "@hiogawa/utils";
 import { CaptionEditor } from "../../components/caption-editor";
 import {
   STORAGE_KEYS,
@@ -6,38 +6,16 @@ import {
   Z_CAPTION_EDITOR_ENTRY_LIST,
   useLocalStorage,
 } from "../../components/caption-editor-utils";
-import { ROUTE_DEF } from "../../misc/routes";
 import { useLoaderDataExtra } from "../../utils/loader-utils";
-import { makeLoader } from "../../utils/loader-utils.server";
 import { ClientOnly } from "../../utils/misc-react";
 import type { PageHandle } from "../../utils/page-handle";
-import { VideoMetadata } from "../../utils/types";
-import { fetchVideoMetadata, parseVideoId } from "../../utils/youtube";
+
+import type { LoaderData } from "./watch.server";
+export { loader } from "./watch.server";
 
 export const handle: PageHandle = {
   navBarTitle: () => "Caption Editor",
 };
-
-//
-// loader
-//
-
-type LoaderData = {
-  videoId: string;
-  videoMetadata: VideoMetadata;
-};
-
-export const loader = makeLoader(async ({ ctx }) => {
-  const query = ROUTE_DEF["/caption-editor/watch"].query.parse(ctx.query);
-  const videoId = parseVideoId(query.v);
-  tinyassert(videoId);
-  const videoMetadata = await fetchVideoMetadata(videoId);
-  return { videoId, videoMetadata } satisfies LoaderData;
-});
-
-//
-// component
-//
 
 export default function Page() {
   return (
